Refactor to put most things into lib
