+++ src/fetch.rs
@@ -1,12 +1,11 @@
-use crate::{Krate, Source};
+use crate::{util, Context, Krate, Source};
 use bytes::{BufMut, Bytes, BytesMut};
 use failure::{bail, Error, ResultExt};
 use log::debug;
 use reqwest::Client;
-use std::{convert::TryFrom, process::Command};
-use tame_gcs::{objects::Object, BucketName, ObjectName};
+use std::process::Command;
+use tame_gcs::objects::Object;
 
-// We just treat versions as opaque strings
 pub fn from_crates_io(client: &Client, krate: &Krate) -> Result<Bytes, Error> {
     match &krate.source {
         Source::CratesIo(_) => {
@@ -16,33 +15,25 @@ pub fn from_crates_io(client: &Client, krate: &Krate) -> Result<Bytes, Error> {
             );
 
             let mut response = client.get(&url).send()?.error_for_status()?;
-            let res = crate::convert_response(&mut response)?;
+            let res = util::convert_response(&mut response)?;
             Ok(res.into_body())
         }
         Source::Git { .. } => via_git(&krate),
     }
 }
 
-pub fn from_gcs(
-    client: &Client,
-    krate: &Krate,
-    bucket: &BucketName<'_>,
-    prefix: &str,
-) -> Result<Bytes, Error> {
-    let object_name = format!("{}{}", prefix, krate.gcs_id());
-    let object_name = ObjectName::try_from(object_name.as_ref())?;
-
-    let dl_req = Object::download(&(bucket, &object_name), None)?;
+pub fn from_gcs(ctx: &Context<'_>, krate: &Krate) -> Result<Bytes, Error> {
+    let dl_req = Object::download(&(&ctx.gcs_bucket, &ctx.object_name(&krate)?), None)?;
 
     let (parts, _) = dl_req.into_parts();
 
     let uri = parts.uri.to_string();
-    let builder = client.get(&uri);
+    let builder = ctx.client.get(&uri);
 
     let request = builder.headers(parts.headers).build()?;
 
-    let mut response = client.execute(request)?.error_for_status()?;
-    let res = crate::convert_response(&mut response)?;
+    let mut response = ctx.client.execute(request)?.error_for_status()?;
+    let res = util::convert_response(&mut response)?;
     Ok(res.into_body())
 }
