+++ src/sync.rs
@@ -1,27 +1,16 @@
+use crate::{fetch, util, Krate, Source};
 use bytes::{Buf, IntoBuf};
-use cargo_fetcher::{fetch, util, Krate, Source};
-use failure::{format_err, Error};
+use failure::Error;
 use log::{error, info};
 use rayon::prelude::*;
-use std::{
-    io::Write,
-    path::{Path, PathBuf},
-};
-
-#[derive(structopt::StructOpt)]
-pub struct Args {
-    /// The root path for cargo. This defaults to either
-    /// CARGO_HOME or HOME/.cargo.
-    #[structopt(short, long, parse(from_os_str))]
-    cache: Option<PathBuf>,
-}
+use std::{io::Write, path::Path};
 
 const INDEX_DIR: &str = "registry/index/github.com-1ecc6299db9ec823";
 const CACHE_DIR: &str = "registry/cache/github.com-1ecc6299db9ec823";
 const SRC_DIR: &str = "registry/src/github.com-1ecc6299db9ec823";
 const GIT_DB_DIR: &str = "git/db";
 
-fn sync_registry_index(ctx: &crate::Context<'_>, root_dir: &Path) -> Result<(), Error> {
+pub fn registry_index(ctx: &crate::Context<'_>, root_dir: &Path) -> Result<(), Error> {
     let index_path = root_dir.join(INDEX_DIR);
 
     // Just skip the index if the git directory already exists,
@@ -39,30 +28,31 @@ fn sync_registry_index(ctx: &crate::Context<'_>, root_dir: &Path) -> Result<(), 
     let krate = Krate {
         name: "crates.io-index".to_owned(),
         version: "1.0.0".to_owned(),
-        source: cargo_fetcher::Source::Git {
+        source: Source::Git {
             url: canonicalized,
             ident,
         },
     };
 
-    let index_data = fetch::from_gcs(&ctx.client, &krate, &ctx.gcs_bucket, ctx.prefix)?;
+    let index_data = fetch::from_gcs(&ctx, &krate)?;
 
     let buf_reader = index_data.into_buf().reader();
     let zstd_decoder = zstd::Decoder::new(buf_reader)?;
 
-    if let Err((_, e)) = cargo_fetcher::unpack_tar(zstd_decoder, index_path) {
+    if let Err((_, e)) = util::unpack_tar(zstd_decoder, index_path) {
         error!("failed to unpack crates.io-index: {}", e);
     }
 
     Ok(())
 }
 
-fn sync_locked_crates(ctx: &crate::Context<'_>, root_dir: &Path) -> Result<(), Error> {
+pub fn locked_crates(ctx: &crate::Context<'_>, root_dir: &Path) -> Result<(), Error> {
     info!("synchronizing {} crates...", ctx.krates.len());
 
     let cache_dir = root_dir.join(CACHE_DIR);
     let src_dir = root_dir.join(SRC_DIR);
     let git_db_dir = root_dir.join(GIT_DB_DIR);
+
     std::fs::create_dir_all(&cache_dir)?;
     std::fs::create_dir_all(&src_dir)?;
     std::fs::create_dir_all(&git_db_dir)?;
@@ -106,7 +96,7 @@ fn sync_locked_crates(ctx: &crate::Context<'_>, root_dir: &Path) -> Result<(), E
     info!("synchronizing {} missing crates...", to_sync.len());
 
     ctx.krates.par_iter().for_each(|krate| {
-        match fetch::from_gcs(&ctx.client, krate, &ctx.gcs_bucket, ctx.prefix) {
+        match fetch::from_gcs(&ctx, krate) {
             Err(e) => error!("failed to download {}: {}", krate, e),
             Ok(krate_data) => {
                 match &krate.source {
@@ -136,7 +126,7 @@ fn sync_locked_crates(ctx: &crate::Context<'_>, root_dir: &Path) -> Result<(), E
 
                         if !src_path.exists() {
                             log::debug!("unpacking {} to src/", krate);
-                            if let Err((_, e)) = cargo_fetcher::unpack_tar(gz_decoder, src_path) {
+                            if let Err((_, e)) = util::unpack_tar(gz_decoder, src_path) {
                                 error!("failed to unpack dependency {}: {}", krate, e);
                             }
                         }
@@ -152,7 +142,7 @@ fn sync_locked_crates(ctx: &crate::Context<'_>, root_dir: &Path) -> Result<(), E
                         };
 
                         let db_path = git_db_dir.join(format!("{}", krate.local_id()));
-                        if let Err((_, e)) = cargo_fetcher::unpack_tar(zstd_decoder, db_path) {
+                        if let Err((_, e)) = util::unpack_tar(zstd_decoder, db_path) {
                             error!("failed to unpack dependency {}: {}", krate, e);
                         }
                     }
@@ -160,58 +150,6 @@ fn sync_locked_crates(ctx: &crate::Context<'_>, root_dir: &Path) -> Result<(), E
             }
         }
     });
-
-    Ok(())
-}
-
-pub fn cmd(ctx: crate::Context<'_>, args: Args) -> Result<(), Error> {
-    let root_dir = args
-        .cache
-        .or_else(|| std::env::var_os("CARGO_HOME").map(PathBuf::from))
-        .or_else(|| dirs::home_dir().map(|hd| hd.join(".cargo")));
-
-    let root_dir = root_dir.ok_or_else(|| format_err!("unable to determine cargo root"))?;
-
-    // There should always be a bin/cargo(.exe) relative to the root directory, at a minimum
-    let cargo_path = {
-        let mut cpath = root_dir.join("bin/cargo");
-
-        if cfg!(target_os = "windows") {
-            cpath.set_extension("exe");
-        }
-
-        cpath
-    };
-
-    if !cargo_path.exists() {
-        return Err(format_err!(
-            "cargo root {} does not seem to contain the cargo binary",
-            root_dir.display()
-        ));
-    }
-
-    // Create the registry directory as it is the root of multiple other ones
-    std::fs::create_dir_all(root_dir.join("registry"))?;
-
-    rayon::join(
-        || {
-            if !ctx.include_index {
-                return;
-            }
-
-            info!("syncing crates.io index");
-            match sync_registry_index(&ctx, &root_dir) {
-                Ok(_) => info!("successfully synced crates.io index"),
-                Err(e) => error!("failed to sync crates.io index: {}", e),
-            }
-        },
-        || match sync_locked_crates(&ctx, &root_dir) {
-            Ok(_) => {
-                info!("finished syncing crates");
-            }
-            Err(e) => error!("failed to sync crates: {}", e),
-        },
-    );
 
     Ok(())
 }
