+++ src/util.rs
@@ -1,6 +1,9 @@
-use failure::Error;
+use failure::{format_err, Error};
 #[allow(deprecated)]
-use std::hash::{Hash, Hasher, SipHasher};
+use std::{
+    hash::{Hash, Hasher, SipHasher},
+    path::{Path, PathBuf},
+};
 use url::Url;
 
 pub fn to_hex(num: u64) -> String {
@@ -98,4 +101,88 @@ pub fn canonicalize_url(url: &Url) -> Result<Url, Error> {
     url.set_query(None);
 
     Ok(url)
+}
+
+pub fn determine_cargo_root(explicit: Option<PathBuf>) -> Result<PathBuf, Error> {
+    let root_dir = explicit
+        .or_else(|| std::env::var_os("CARGO_HOME").map(PathBuf::from))
+        .or_else(|| dirs::home_dir().map(|hd| hd.join(".cargo")));
+
+    let root_dir = root_dir.ok_or_else(|| format_err!("unable to determine cargo root"))?;
+
+    // There should always be a bin/cargo(.exe) relative to the root directory, at a minimum
+    // there are probably ways to have setups where this doesn't hold true, but this is simple
+    // and can be fixed later
+    let cargo_path = {
+        let mut cpath = root_dir.join("bin/cargo");
+
+        if cfg!(target_os = "windows") {
+            cpath.set_extension("exe");
+        }
+
+        cpath
+    };
+
+    if !cargo_path.exists() {
+        return Err(format_err!(
+            "cargo root {} does not seem to contain the cargo binary",
+            root_dir.display()
+        ));
+    }
+
+    Ok(root_dir)
+}
+
+pub fn convert_response(
+    res: &mut reqwest::Response,
+) -> Result<tame_gcs::http::Response<bytes::Bytes>, Error> {
+    use bytes::BufMut;
+
+    let body = bytes::BytesMut::with_capacity(res.content_length().unwrap_or(4 * 1024) as usize);
+    let mut writer = body.writer();
+    res.copy_to(&mut writer)?;
+    let body = writer.into_inner();
+
+    let mut builder = tame_gcs::http::Response::builder();
+
+    builder.status(res.status()).version(res.version());
+
+    let headers = builder
+        .headers_mut()
+        .ok_or_else(|| format_err!("failed to convert response headers"))?;
+
+    headers.extend(
+        res.headers()
+            .into_iter()
+            .map(|(k, v)| (k.clone(), v.clone())),
+    );
+
+    Ok(builder.body(body.freeze())?)
+}
+
+pub fn unpack_tar<R: std::io::Read, P: AsRef<Path>>(stream: R, dir: P) -> Result<R, (R, Error)> {
+    let mut archive_reader = tar::Archive::new(stream);
+
+    let dir = dir.as_ref();
+
+    if let Err(e) = archive_reader.unpack(dir) {
+        // Attempt to remove anything that may have been written so that we
+        // _hopefully_ don't actually mess up cargo
+        if dir.exists() {
+            if let Err(e) = remove_dir_all::remove_dir_all(dir) {
+                log::error!(
+                    "error trying to remove contents of {}: {}",
+                    dir.display(),
+                    e
+                );
+            }
+        }
+
+        return Err((
+            archive_reader.into_inner(),
+            format_err!("failed to unpack: {}", e),
+        ));
+    }
+
+    Ok(archive_reader.into_inner())
 }
